Fixed the move system Added a DestinationComponent Changed it so the texture positions are at the center of the textures Zombies now go to where the click is and stop there Made everything related to the move system floats due to rounding errors we were getting when they were ints

+++ TheZGame/core/src/com/zgame/ui/IClickHandler.java
@@ -2,6 +2,6 @@ package com.zgame.ui;
 
 public interface IClickHandler {
 
-	public boolean processInput(int x, int y);
+	public boolean processInput(float x, float y);
 	
 }
